!BREAKING: [hub] 허브, 허브라우트, 맵 서비스를 허브 서비스로 통합 - 허브라우트와 맵 서비스의 기능을 허브 서비스로 통합 - 기존의 별도 허브라우트 및 맵 마이크로서비스 제거 - 모든 관련 작업을 처리하도록 허브 서비스 업데이트 - 의존성 및 설정 조정 - 새로운 구조를 반영하여 API 엔드포인트 업데이트
