+++ hub/src/main/java/com/sparta/hub/application/configuration/jpa/JpaConfig.java
@@ -1,0 +1,23 @@
+package com.sparta.hub.application.configuration.jpa;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import com.sparta.hub.application.configuration.auditing.AuditAwareImpl;
+import jakarta.persistence.EntityManager;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+    @Bean
+    public AuditorAware<String> auditorProvider() {
+        return new AuditAwareImpl();
+    }
+
+    @Bean
+    public JPAQueryFactory jpaQueryFactory(EntityManager entityManager) {
+        return new JPAQueryFactory(entityManager);
+    }
+}
