+++ hub/src/main/java/com/sparta/hub/domain/map/model/Coordinates.java
@@ -1,0 +1,11 @@
+package com.sparta.hub.domain.map.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class Coordinates {
+    private final double latitude;
+    private final double longitude;
+}
