--- hub/src/main/java/com/sparta/hub/infrastructure/configuration/cache/CacheConfig.java
@@ -1,52 +1,0 @@
-package com.sparta.hub.infrastructure.configuration.cache;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.sparta.hub.domain.model.Hub;
-import java.time.Duration;
-
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.cache.CacheKeyPrefix;
-import org.springframework.data.redis.cache.RedisCacheConfiguration;
-import org.springframework.data.redis.cache.RedisCacheManager;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.RedisSerializationContext.SerializationPair;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-@Configuration
-@EnableCaching
-public class CacheConfig {
-
-    @Bean
-    public ObjectMapper redisObjectMapper() {
-        ObjectMapper objectMapper = new ObjectMapper();
-        objectMapper.registerModule(new JavaTimeModule());
-        objectMapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-        objectMapper.disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
-
-        return objectMapper;
-    }
-
-    @Bean
-    public RedisCacheManager cacheManager(RedisConnectionFactory redisConnectionFactory, ObjectMapper redisObjectMapper) {
-        // Hub 타입에 대한 직렬화를 명시적으로 설정
-        Jackson2JsonRedisSerializer<Hub> serializer = new Jackson2JsonRedisSerializer<>(redisObjectMapper, Hub.class);
-
-        RedisCacheConfiguration defaultCacheConfig = RedisCacheConfiguration
-                .defaultCacheConfig()
-                .disableCachingNullValues()
-                .entryTtl(Duration.ofDays(1))
-                .computePrefixWith(CacheKeyPrefix.simple())
-                .serializeKeysWith(SerializationPair.fromSerializer(new StringRedisSerializer())) // 키 직렬화
-                .serializeValuesWith(SerializationPair.fromSerializer(serializer)); // 값 직렬화
-
-        return RedisCacheManager.builder(redisConnectionFactory)
-                .cacheDefaults(defaultCacheConfig)
-                .build();
-    }
-}
