--- hub/src/main/java/com/sparta/hub/presentation/dto/request/HubRequest.java
@@ -1,17 +1,0 @@
-package com.sparta.hub.presentation.dto.request;
-
-import com.sparta.hub.application.dto.HubDto;
-import lombok.Getter;
-
-@Getter
-public class HubRequest {
-
-    private String name;
-    private String address;
-
-    public HubDto toDTO() {
-        return HubDto.createWithoutManager(name, address);
-    }
-}
-
-
