+++ hub/src/main/resources/application.yml
@@ -13,10 +13,6 @@ spring:
       ddl-auto: update
     show-sql: true
 
-  cache:
-    redis:
-      time-to-live: ${REDIS_TTL:600000}
-
   data:
     redis:
       host: ${REDIS_HOST:localhost}
@@ -43,5 +39,8 @@ management:
 logging:
   level:
     root: INFO
-    org:
-      springframework: DEBUG
+
+google:
+  maps:
+    api:
+      key: ${YOUR_GOOGLE_MAPS_API_KEY}
